towards-targets: Move DLE inside newly created targets
