--- devito/dle/__init__.py
@@ -1,5 +1,0 @@
-from devito.dle.blocking_utils import *  # noqa
-from devito.dle.parallelizer import (NThreads, NThreadsNested,  # noqa
-                                     NThreadsNonaffine, Ompizer, OmpizerGPU)
-from devito.dle.rewriters import *  # noqa
-from devito.dle.transformer import *  # noqa
