+++ devito/ops/__init__.py
@@ -4,8 +4,8 @@ JIT-compile, and run kernels on multiple architectures.
 """
 
 from devito.archinfo import Cpu64
-from devito.dle import PlatformRewriter, modes
 from devito.parameters import Parameters, add_sub_configuration
+from devito.targets import PlatformRewriter, modes
 
 # Add OPS-specific DLE modes
 modes.add(Cpu64, {'noop': PlatformRewriter,
