+++ devito/targets/__init__.py
@@ -1,0 +1,1 @@
+from .dle import *  # noqa
