+++ devito/targets/dle/__init__.py
@@ -1,0 +1,5 @@
+from .blocking_utils import *  # noqa
+from .parallelizer import (NThreads, NThreadsNested,  # noqa
+                           NThreadsNonaffine, Ompizer, OmpizerGPU)
+from .rewriters import *  # noqa
+from .transformer import *  # noqa
