+++ devito/targets/dle/rewriters.py
@@ -5,8 +5,6 @@ from time import time
 
 import cgen
 
-from devito.dle.blocking_utils import Blocker, BlockDimension
-from devito.dle.parallelizer import Ompizer, OmpizerGPU
 from devito.exceptions import DLEException
 from devito.ir.iet import (Call, Iteration, List, HaloSpot, Prodder, PARALLEL, VECTOR,
                            FindSymbols, FindNodes, FindAdjacent, MapNodes, Transformer,
@@ -14,6 +12,8 @@ from devito.ir.iet import (Call, Iteration, List, HaloSpot, Prodder, PARALLEL, V
 from devito.logger import perf_adv, dle_warning as warning
 from devito.mpi import HaloExchangeBuilder, HaloScheme
 from devito.parameters import configuration
+from devito.targets.dle.blocking_utils import Blocker, BlockDimension
+from devito.targets.dle.parallelizer import Ompizer, OmpizerGPU
 from devito.tools import DAG, as_tuple, filter_ordered, generator
 
 __all__ = ['PlatformRewriter', 'CPU64Rewriter', 'Intel64Rewriter', 'PowerRewriter',
